extract CUDA-independent parts of aten harness
Put checkRtol into test_harness_aten.h as it does not depend on CUDA and
can be used transparently with CUDA and CPU tensors.
